+++ src/main/java/com/owlmaddie/goals/FleePlayerGoal.java
@@ -1,8 +1,9 @@
 package com.owlmaddie.goals;
 
-import net.minecraft.entity.ai.goal.Goal;
-import net.minecraft.entity.ai.pathing.EntityNavigation;
+import net.minecraft.entity.ai.FuzzyTargeting;
+import net.minecraft.entity.ai.pathing.Path;
 import net.minecraft.entity.mob.MobEntity;
+import net.minecraft.entity.mob.PathAwareEntity;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.math.Vec3d;
 
@@ -10,47 +11,47 @@ import java.util.EnumSet;
 
 /**
  * The {@code FleePlayerGoal} class instructs a Mob Entity to flee from the current player
- * and only recalculates the flee path when it has reached its destination and the player is close again.
+ * and only recalculates path when it has reached its destination and the player is close again.
  */
-public class FleePlayerGoal extends Goal {
+public class FleePlayerGoal extends PlayerBaseGoal {
     private final MobEntity entity;
-    private ServerPlayerEntity targetPlayer;
-    private final EntityNavigation navigation;
     private final double speed;
     private final float fleeDistance;
 
     public FleePlayerGoal(ServerPlayerEntity player, MobEntity entity, double speed, float fleeDistance) {
-        this.targetPlayer = player;
+        super(player);
         this.entity = entity;
         this.speed = speed;
         this.fleeDistance = fleeDistance;
-        this.navigation = entity.getNavigation();
         this.setControls(EnumSet.of(Control.MOVE));
     }
 
     @Override
     public boolean canStart() {
-        return this.targetPlayer != null && this.entity.squaredDistanceTo(this.targetPlayer) < fleeDistance * fleeDistance;
+        return super.canStart() && this.entity.squaredDistanceTo(this.targetEntity) < fleeDistance * fleeDistance;
     }
 
     @Override
     public boolean shouldContinue() {
-        return this.navigation.isFollowingPath();
+        return super.canStart() && this.entity.squaredDistanceTo(this.targetEntity) < fleeDistance * fleeDistance;
     }
 
     @Override
     public void stop() {
-        this.navigation.stop();
+        this.entity.getNavigation().stop();
     }
 
     private void fleeFromPlayer() {
-        Vec3d fleeDirection = new Vec3d(
-                this.entity.getX() - this.targetPlayer.getX(),
-                this.entity.getY() - this.targetPlayer.getY(),
-                this.entity.getZ() - this.targetPlayer.getZ()
-        ).normalize();
-        Vec3d fleeTarget = fleeDirection.multiply(fleeDistance).add(this.entity.getPos());
-        this.navigation.startMovingTo(fleeTarget.x, fleeTarget.y, fleeTarget.z, this.speed);
+        int roundedFleeDistance = Math.round(fleeDistance);
+        Vec3d fleeTarget = FuzzyTargeting.findFrom((PathAwareEntity)this.entity, roundedFleeDistance,
+                roundedFleeDistance, this.targetEntity.getPos());
+
+        if (fleeTarget != null) {
+            Path path = this.entity.getNavigation().findPathTo(fleeTarget.x, fleeTarget.y, fleeTarget.z, 0);
+            if (path != null) {
+                this.entity.getNavigation().startMovingAlong(path, this.speed);
+            }
+        }
     }
 
     @Override
@@ -60,9 +61,7 @@ public class FleePlayerGoal extends Goal {
 
     @Override
     public void tick() {
-        // Only recalculate the flee path if the entity has reached its destination or doesn't have an active path,
-        // and the player is within the flee distance again.
-        if (!this.navigation.isFollowingPath() && this.entity.squaredDistanceTo(this.targetPlayer) < fleeDistance * fleeDistance) {
+        if (!this.entity.getNavigation().isFollowingPath()) {
             fleeFromPlayer();
         }
     }