+++ src/main/java/com/owlmaddie/goals/FollowPlayerGoal.java
@@ -1,24 +1,24 @@
 package com.owlmaddie.goals;
 
 import com.owlmaddie.controls.LookControls;
-import net.minecraft.entity.ai.goal.Goal;
+import net.minecraft.entity.ai.FuzzyTargeting;
 import net.minecraft.entity.ai.pathing.EntityNavigation;
-import net.minecraft.entity.mob.MobEntity;
+import net.minecraft.entity.mob.*;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.util.math.Vec3d;
 
 import java.util.EnumSet;
 
 /**
- * The {@code FollowPlayerGoal} class instructs a Mob Entity to follow the current player.
+ * The {@code FollowPlayerGoal} class instructs a Mob Entity to follow the current target entity.
  */
-public class FollowPlayerGoal extends Goal {
+public class FollowPlayerGoal extends PlayerBaseGoal {
     private final MobEntity entity;
-    private ServerPlayerEntity targetPlayer;
     private final EntityNavigation navigation;
     private final double speed;
 
     public FollowPlayerGoal(ServerPlayerEntity player, MobEntity entity, double speed) {
-        this.targetPlayer = player;
+        super(player);
         this.entity = entity;
         this.speed = speed;
         this.navigation = entity.getNavigation();
@@ -28,13 +28,13 @@ public class FollowPlayerGoal extends Goal {
     @Override
     public boolean canStart() {
         // Start only if the target player is more than 8 blocks away
-        return this.targetPlayer != null && this.entity.squaredDistanceTo(this.targetPlayer) > 64;
+        return super.canStart() && this.entity.squaredDistanceTo(this.targetEntity) > 64;
     }
 
     @Override
     public boolean shouldContinue() {
-        // Continue unless the entity gets within 3.x blocks of the player
-        return this.targetPlayer != null && this.targetPlayer.isAlive() && this.entity.squaredDistanceTo(this.targetPlayer) > 12;
+        // Continue unless the entity gets within 3 blocks of the player
+        return super.canStart() && this.entity.squaredDistanceTo(this.targetEntity) > 9;
     }
 
     @Override
@@ -45,8 +45,24 @@ public class FollowPlayerGoal extends Goal {
 
     @Override
     public void tick() {
-        // Look at the player and start moving towards them
-        LookControls.lookAtPlayer(this.targetPlayer, this.entity);
-        this.navigation.startMovingTo(this.targetPlayer, this.speed);
+        if (this.entity instanceof EndermanEntity || this.entity instanceof EndermiteEntity || this.entity instanceof ShulkerEntity) {
+            // Certain entities should teleport to the player if they get too far
+            if (this.entity.squaredDistanceTo(this.targetEntity) > 256) {
+                Vec3d targetPos = findTeleportPosition(12);
+                if (targetPos != null) {
+                    this.entity.teleport(targetPos.x, targetPos.y, targetPos.z);
+                }
+            }
+        } else {
+            // Look at the player and start moving towards them
+            if (this.targetEntity instanceof ServerPlayerEntity) {
+                LookControls.lookAtPlayer((ServerPlayerEntity)this.targetEntity, this.entity);
+            }
+            this.navigation.startMovingTo(this.targetEntity, this.speed);
+        }
+    }
+
+    private Vec3d findTeleportPosition(int distance) {
+        return FuzzyTargeting.findTo((PathAwareEntity)this.entity, distance, distance, this.targetEntity.getPos());
     }
 }